RxBluetoothKit and Hashable conformance
Added extensions for Characteristic and Peripheral. They both
now conform to Hashable protocol nad provide hash values for
themselves.
