+++ ExampleApp/ExampleApp/Common/Characteristic/Characterist+CBCharacteristicWriteType.swift
@@ -4,6 +4,7 @@ import CoreBluetooth
 
 extension Characteristic {
 
+    // Utility method, which helps
     func determineWriteType() -> CBCharacteristicWriteType? {
         let writeType =  self.properties.contains(.write) ? CBCharacteristicWriteType.withResponse :
                 self.properties.contains(.writeWithoutResponse) ? CBCharacteristicWriteType.withResponse : nil
