+++ ExampleApp/ExampleApp/Common/Services/RxBluetoothKitService.swift
@@ -130,25 +130,3 @@ final class RxBluetoothKitService {
     }
 
 }
-
-extension Peripheral: Hashable {
-
-    // DJB Hashing
-    public var hashValue: Int {
-        let scalarArray: [UInt32] = []
-        return scalarArray.reduce(5381) {
-            ($0 << 5) &+ $0 &+ Int($1)
-        }
-    }
-}
-
-extension Characteristic: Hashable {
-
-    // DJB Hashing
-    public var hashValue: Int {
-        let scalarArray: [UInt32] = []
-        return scalarArray.reduce(5381) {
-            ($0 << 5) &+ $0 &+ Int($1)
-        }
-    }
-}
